+++ changelog.md
@@ -2,11 +2,10 @@
 
 ## master
 
-### misc
-
 - changed keywords for npm package
 - introduced changelog
 - improved readme
+- make templates exchangeable
 
 ## 0.0.1